CL784864 BSLRM865-238: capture: vvcam: Add the autoexposure ROI set support
If set hist/exp roi when isp streamming, mean luma is not correct.
The reason is that hist/exp has no shadow register, and need to add
config signal to enable the settings to be configured.

So need to set hist/exp when frame done signal is received, and add
cfg_upd signal after setting roi relate register.

Signed-off-by: Alice Yuan <alice.yuan@nxp.com>
Reviewed-by: Robby Cai <robby.cai@nxp.com>